Explicitly fontify before indenting region
In the Emacs Erlang mode, we depend on font-lock-syntactic-keywords to
recognise whether dollar signs before single and double quotes are
character constants or part of a string or atom.  Since Emacs performs
syntax highlighting lazily, indenting parts of a buffer that haven't
been displayed yet (such as when indenting an entire buffer, or running
the "test_indentation" target of the makefile) would behave incorrectly
when faced with such strings.

This change makes erlang-indent-region call font-lock-fontify-region
before indenting, and modifies test.erl.orig and test.erl.indented to
test for this behaviour.