Improve handling of critical extensions in SSL certificates
If certificate validation fails because of a critical extension that
wasn't understood, log this with a 'unknown_critical_extension'
message, including the extension itself and the subject and issuer of
the certificate.

In verify_none mode, pretend that we "understand" critical extensions,
logging them in an 'ignoring_critical_extension' message.