Add -start_epmd command line option
Add a command line option that lets you disable automatic starting of
epmd when starting a distributed node.

This differs from the undocumented setting -no_epmd, in that it does
not affect the starting of an erl_epmd process within
erl_distribution: the newly started node will expect an epmd instance
to have been started previously.